Forcebly sets utf8 encoding on windows
As the result of defaultTest expects to be
able to write utf-8 to stdout/stderr, we
ensure that their encoding is set accordingly.

This should fix failures like:
```
cardano-crypto-golden-tests.exe: <stdout>: commitBuffer: invalid
argument (invalid character)
```

Note the addition of the Encoding module. This is
necessary as multi-line strings (in GoldenTests.hs)
do not mix well with CPP.