cbits: Squash a Wstringop-overread error
Pretty sure this warning is only generated with GCC versions >= 14.2.

A type and a function were defined as follows:

    typedef unsigned char ed25519_secret_key[64];
    void cardano_crypto_ed25519_publickey(const ed25519_secret_key sk, ed25519_public_key pk);

and then the code had:

    uint8_t zl8[32];
    cardano_crypto_ed25519_publickey(zl8, pub_zl8);

The compiler error was squashed by changing the type of `uint8_t [32]` to
`ed25519_secret_key` which is 64 bytes in size and then updating the
`memset` call to zero the whole array.