Bump version to 1.3.0
There's already a 1.2.0 version on CHaP, so in order to publish
cardano-crypto on Hackaage (which is needed for cardano-addresses to
be published on Hackage), we want to prevent potential conflicts to
happen.

This also removes dependency on CHaP from the `cabal.project` to
guarantee the project can be built with dependencies drawn only from
Hackage.